requirements.txt
mimesis==10.0.0
numpy==1.24.3
pandas==2.0.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
